manually crafted patch for SI-6188 (ICodeReader skips exception handlers)
A backport of https://github.com/magarciaEPFL/scala/commit/61cc8ff61c81a1276a921ad5288ee3bebea1c96e

A test case for it can't rely on scala.util.Success which as of the checkout isn't there yet. But given it'll be added in another backport, seems like it can wait till then. The test case is tricky in that it requires a previously compiled classfile (where bytecode for an @inline method with exception handlers can be found). Ticket SI-6188 contains a self-contained example (not dependent on scala.util.Success) with _1 and _2 .scala files . Alternatively that test case could be used instead.

The fix above for SI-6188 is safe in the sense that it prevents those methods (ie read by ICodeReader containing exception handlers) from being inlined, even if marked @inline. It's suboptimal in that sense, the fix for SI-6157 takes care of checking additional conditions and performing "merging" of exception handler tables, such that those methods are inlined. But I guess SI-6157 has not been backported, and is in fact not terribly necessary. Just mentioning this for completeness.