TODO pending test waiting for support to be added for "really static" inlining
With -opimise , code is emitted (by GenASM) that expects an instance method where a really-static now exists, the runtime error in this case is:
   java.lang.IncompatibleClassChangeError: Expecting non-static method scala.tools.partest.instrumented.Instrumentation$.startProfiling()V

With -neo:o3, callsites are recorded by a javaagent. As of now inlining of methods in really-static static modules (e.g., MyPredef) isn't yet supported, thus the recording shows callsites that will go away after that support arrives.