statification, @enforceStatic module-classes and their usages-sites
"Statification" is a transformation performed on-the-fly by GenBCode
for "really static" module classes and their usage sites.

An @enforceStatic module-classes has all of its members (for example,
extension methods of a custom value classes) emitted bytecode-level
static, EXCEPT those members overriding methods in the Object API.
 reason for this is that some clients may invoke those methods on
the singleton (viewed as an object) and expect the
overriding implementation to be dispatched.

Additionally, @enforceStatic also results in invokestatic callsites
for what used to be invokevirtual callsites with the singleton
as receiver, ie the singleton is not loaded anymore before
any arguments are loaded.

There's an interplay with @inline methods in an
@enforceStatic module-class: the inliner gets permission to have
an `invokestatic` (targeting the method in question) replaced with
its body, thus potentially skipping running the class-initializer
for the associated module-class.