Level 1 intra-method optimiz: Copy propagation
Intra-method optimizations are performed until a fixpoint is reached.

"Copy propagation" replaces the last link in a chain of data accesses
by a direct access to the chain-start.

A "chain of data accesses" refers to a assignments of the form shown below,
without any intervening rewriting of v, v1, v2, ..., v9:

      v1 =  v
       ...
      v2 = v1
       ...
      v9 = v8

After CopyPropagator.transform has run,
`LOAD v9` has been replaced with `LOAD v`.
Similarly for `LOAD v1` to `LOAD v8`.

Other than that, this method changes nothing else:
in particular STORE v1` to `STORE v9` are left as-is.
Removing dead-stores is the job of `DeadStoreElim`.