GenBCode: test built-in jump-chains collapsing
Code that contains redundant jumps usually arrives to both
GenASM and GenBCode, for a variety of reasons. To name a few:

  - test/files/run/private-inline.scala after -optimize

  - virtual pattern matcher, example below

The backend(s) manage to get rid of the redundant jumps, for example:

  - GenASM, using collapseJumpOnlyBlocks()

  - GenBCode, using JumpChainsCollapser

The test case in this commit shows how GenBCode reduces
the number of GOTOs from 22 down to 7.