phase assembly becomes backend-dependent, and -Xshow-phases shows that
"Phase assembly" refers to what `computeInternalPhases()` does.
Before this commit, the resulting set of standard phases was constant,
irrespective of the optimizations in effect.
In particular, before this commit all optimization phases were always
included by `computeInternalPhases()`.

With the addition of `GenBCode`, we take the opportunity to follow
a more clean approach: assembling only those phases
that are going to be actually used.

- in case of using `GenBCode`,
  that means leaving out GenICode, the old optimizer, and GenASM.
- in case `GenASM` is chosen,
  the standard phases are assembled as before.

Setting `-Xshow-phases` therefore produces the same output "as before"
unless `-Ybackend:GenBCode` is chosen, in which case
phases like `GenICode` and `GenASM` aren't shown.

Regarding tests that employ `-Xshow-phases`
(for example, in connection with compiler plugins)
it's enough to have those tests run with the default backend
(ie `GenASM`) unless some aspect specific to `GenBCode` phase ordering
is to be tested.