styling dclosure endpoints for LambdaMetaFactory
An aspect relevant for future compatibility with JDK 8.

A LambdaMetaFactory in JDK8 delivers lambdas that assume
the (hoisted) target method to have arguments for all captured values
and for apply-args (in that order).

In contrast, "hoisted target methods" emerge from LambdaLift with
formal-params for captured values appended after those for apply-args.

Moreover, for mixins, the first formal of an endpoint in
an implementation class denotes the self-reference.

This commit makes dclosure-endpoints ready for LambdaMetaFactory.