GenBCode debug utility: Type-Flow Analysis
Type-Flow Analysis (TFA) is an iterative dataflow-analysis
that usually finds more precise types for local-vars and stack-slots
than those declared statically.

It can also be used to detect some forms of invalid bytecode,
for example to detect "different stack heights on branch merging".

In BCodeOpt, neither the method-inliner nor closure-inliner
rely on TFA because it's enough to rely on at-inline annotations
and the static type of the receiver (at a callsite) to benefit from inlining
(to recap, features like: only at-inline methods are inlined,
and inlining is explicitly chosen by the developer).