more precise abstract-values in ASM dataflow-analysis
Before this commit, there was only one catch-all mechanism to
create the abstract values of a custom dataflow analyses (DFAs),
namely via Interpreter.newValue(Type)

That state of affairs leads to imprecission, as can be seen
for example when creating a nullness-tracking analysis
(an instance method should have its this variable initially
tracked as non-null).

This commit adds methods to asm.tree.analysis.Analyzer to
allow overridding what abstract values are of interest for:
  - formal params in a method
  - local vars in a method
  - exception on the operand stack on entry to an
    exception-handling block

Client classes that use asm.tree.analysis.Analyzer continue to work,
and need not be recompiled. The additional methods, unless overridden,
( `newFormal()` , `newFormalLocal()` , and `newException()` )
delegate to `Interpreter.newValue()` which is the previous behavior.